+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to `when` are documented here.
 ## 0.2.0
 
 - Fixed an issue where admin code was same as country code.
+- Add support for colors.
+- Improved error message output.
+- Added week day output.
 
 ## 0.1.0